refactor: enhance framework structure and error handling, update descriptions
