+++ src/constants/decorators.contants.ts
@@ -1,1 +1,4 @@
-export const ZEN_MODULE_METADATA = 'zen:module';
+export const ZEN_MODULE_METADATA = Symbol.for('zen:module');
+export const ZEN_MIDDLEWARE_METADATA = Symbol.for('zen:middleware');
+export const ZEN_CONTROLLER_BASE_PATH_METADATA = Symbol.for('zen:controller:basePath');
+export const ZEN_CONTROLLER_ROUTES_METADATA = Symbol.for('zen:controller:routes');
